feat(user): add user policy and its functional tests

+++ app/controllers/v1/users_controller.ts
@@ -1,4 +1,4 @@
-import User from '#models/user';
+import UserPolicy from '#policies/user_policy';
 import UserService, { UserAttributes } from '#services/user_service';
 import { createUserValidator, updateUserValidator } from '#validators/user_validator';
 import { inject } from '@adonisjs/core';
@@ -11,14 +11,22 @@ export default class UsersController {
   /**
    * Display a list of resource
    */
-  async index({ response }: HttpContext) {
+  async index({ bouncer, response }: HttpContext) {
+    if (await bouncer.with(UserPolicy).denies('list')) {
+      return response.notFound();
+    }
+
     return response.ok(await this.$service.list());
   }
 
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response }: HttpContext) {
+  async store({ bouncer, request, response }: HttpContext) {
+    if (await bouncer.with(UserPolicy).denies('create')) {
+      return response.forbidden();
+    }
+
     const attributes: Partial<UserAttributes> = await request.validateUsing(createUserValidator);
     return response.created(await this.$service.store(attributes as UserAttributes));
   }
@@ -26,14 +34,22 @@ export default class UsersController {
   /**
    * Show individual record
    */
-  async show({ params, response }: HttpContext) {
+  async show({ bouncer, params, response }: HttpContext) {
+    if (await bouncer.with(UserPolicy).denies('show')) {
+      return response.notFound();
+    }
+
     return response.ok(await this.$service.findOrFail(params.id));
   }
 
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ bouncer, params, request, response }: HttpContext) {
+    if (await bouncer.with(UserPolicy).denies('update')) {
+      return response.forbidden();
+    }
+
     const attributes: Partial<UserAttributes> = await request.validateUsing(
       updateUserValidator(params.id)
     );
@@ -43,10 +59,12 @@ export default class UsersController {
   /**
    * Soft delete record
    */
-  async archive({ params, response }: HttpContext) {
-    const user: User = await this.$service.findOrFail(params.id);
+  async archive({ bouncer, params, response }: HttpContext) {
+    if (await bouncer.with(UserPolicy).denies('archive')) {
+      return response.forbidden();
+    }
 
-    await this.$service.archive(user);
+    await this.$service.archive(params.id);
 
     return response.noContent();
   }
@@ -54,7 +72,11 @@ export default class UsersController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ bouncer, params, response }: HttpContext) {
+    if (await bouncer.with(UserPolicy).denies('destroy')) {
+      return response.forbidden();
+    }
+
     await this.$service.delete(params.id);
 
     return response.noContent();
